Refactoring of trace logging, step 1.  Custom get_logger that returns a Logger with the extra TRACE functionality.
