+++ lib/galaxy/__init__.py
@@ -18,16 +18,3 @@ if not hasattr(logging, 'NullHandler'):
         def emit(self, record):
             pass
     logging.NullHandler = NullHandler
-
-
-# Add custom "TRACE" log level for ludicrous verbosity.
-LOGLV_TRACE = 5
-logging.addLevelName(LOGLV_TRACE, "TRACE")
-
-
-def trace(self, message, *args, **kws):
-    if self.isEnabledFor(LOGLV_TRACE):
-        self._log(LOGLV_TRACE, message, args, **kws)
-
-
-logging.Logger.trace = trace
