+++ lib/galaxy/util/log/__init__.py
@@ -1,3 +1,7 @@
+import logging
+import types
+
+
 class TraceLogger(object):
 
     def __init__(self, name):
@@ -5,3 +9,19 @@ class TraceLogger(object):
 
     def log(**kwargs):
         raise TypeError("Abstract Method")
+
+
+# Add custom "TRACE" log level for ludicrous verbosity.
+LOGLV_TRACE = 5
+logging.addLevelName(LOGLV_TRACE, "TRACE")
+
+
+def trace(self, message, *args, **kws):
+    if self.isEnabledFor(LOGLV_TRACE):
+        self._log(LOGLV_TRACE, message, args, **kws)
+
+
+def get_logger(name=None):
+    logger = logging.getLogger(name)
+    logger.trace = types.MethodType(trace, logger)
+    return logger
